refactor: reorganize components location to find them more easier
