+++ src/app/(main)/dashboard/components/forms/reception-form/serial-selector.tsx
@@ -18,10 +18,10 @@ import ModalForm from '@/modules/common/components/modal-form'
 import { useCallback, useEffect, useState, useTransition } from 'react'
 import { Loader2, MousePointerClickIcon } from 'lucide-react'
 import { Button } from '@/modules/common/components/button'
-import { useSelectedItemCardContext } from './context/card-context'
+import { useSelectedItemCardContext } from '../../../../../../lib/context/selected-item-card-context'
 import { SerialWithRenglon } from '@/types/types'
 import { DataTable } from '@/modules/common/components/table/data-table'
-import { receptionSerialColumns } from './serial-columns'
+import { receptionSerialColumns } from '../../columns/serial-selector-columns'
 import { NumericFormat } from 'react-number-format'
 import { getSerialsByItemEnabled } from '@/lib/actions/serials'
