+++ src/app/(main)/dashboard/components/selected-item-card-header/index.tsx
@@ -19,7 +19,7 @@ import {
   AlertDialogTrigger,
 } from '@/modules/common/components/alert-dialog'
 import { AlertDialogImage } from '@/modules/common/components/alert-dialog/alert-dialog-image'
-import { useSelectedItemCardContext } from '../forms/reception-form/context/card-context'
+import { useSelectedItemCardContext } from '../../../../../lib/context/selected-item-card-context'
 
 export const SelectedItemCardHeader = () => {
   const { itemData, removeCard, isEditing } = useSelectedItemCardContext()
